+++ packages/winnow/CHANGELOG.md
@@ -2,6 +2,10 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## Unreleased
+### Added
+* Add support for Point geometry filters
+
 ## [2.3.0] - 03-31-2021
 ### Added
 * Type definition file for TypeScript